+++ docs/release_notes.rst
@@ -8,3 +8,4 @@ Substeps Runner - Release Notes
 - Changes as a result of core changes to Notifications.
 - BUG: failures in @BeforeAllFeatures not failing the maven build
 - Doc changes clarified use of tags
+- Report builder no longer essential